+++ src/Index/SourceFileChange.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Lsp\Index;
+
+enum SourceFileChange
+{
+    case Changed;
+    case Unchanged;
+    case Untracked;
+}